+++ plugins/commands/up/start_mixins.rb
@@ -20,6 +20,18 @@ module VagrantPlugins
           options[:provision_types] = list.map { |type| type.to_sym }
         end
       end
+
+      # This validates the provisioner flags and raises an exception
+      # if there are invalid ones.
+      def validate_provisioner_flags!(options)
+        (options[:provision_types] || []).each do |type|
+            klass = Vagrant.plugin("2").manager.provisioners[type]
+            if !klass
+              raise Vagrant::Errors::ProvisionerFlagInvalid,
+                name: type.to_s
+            end
+        end
+      end
     end
   end
 end